requirements.txt
requests
google-genai
openai
nltk
panda
scikit-learn
numpy #required for scikit-learn
scipy #required for scikit-learn